Finished UnreliableSequenced implementation
NOTE: There can only be 1 channel per connection/user for now

Everytime a message is enqueued in the ConcurrentCircularQueue and the channel is UnreliableSequenced, we check the sequence number of all queued messages and compare them with the new one.
If the new one is smaller than an existing one, we insert it in the right position and move the more recent ones forward.

To make this work, a lock needs to be used in the ConcurrentCircularQueue when sorting.

Another required change I had to make was to move the _userEventQueue to the Connection class because the sequence numbers need to be per connection and not global (1 UnreliableSequenced channel per connection is supported)